Improved the loop and other changes.
The loop now creates a table and put each cell element into an array. It
also has enough cells for the longest month possibly. Found a formula to
calculate the day of the week. Created a the dayObj object for later
use.  Added a seperate <table> for the days of the week names.